requirements.txt
beautifulsoup4==4.6.0
bs4==0.0.1
certifi==2017.11.5
chardet==3.0.4
click==6.7
Flask==0.12.1
future==0.16.0
git+https://gitlab.com/graphviz/graphviz.git
gunicorn==19.7.1
idna==2.6
itsdangerous==0.24
jieba==0.39
Jinja2==2.10
MarkupSafe==1.0
python-telegram-bot==5.3.0
requests==2.18.4
six==1.11.0
transitions==0.5.0
urllib3==1.22
Werkzeug==0.13
git+https://github.com/pygraphviz/pygraphviz.git
